Move Work component inside page itself
